drinol/accounts/serializers.py
from rest_framework import serializers

from drinol.accounts.models import CustomUser


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = CustomUser
        fields = (
            "id",
            "username",
        )
